combine api parsing python files
